// components/generated/wundergraph.internal.client.ts
// Code generated by wunderctl. DO NOT EDIT.

import {
	AddMessageResponse,
	AddMessageInput,
	InternalAddMessageInput,
	InjectedAddMessageInput,
	ChangeUserNameResponse,
	ChangeUserNameInput,
	InternalChangeUserNameInput,
	InjectedChangeUserNameInput,
	DeleteAllMessagesByUserEmailResponse,
	DeleteAllMessagesByUserEmailInput,
	InternalDeleteAllMessagesByUserEmailInput,
	InjectedDeleteAllMessagesByUserEmailInput,
	HelloResponse,
	MessagesResponse,
	MockQueryResponse,
	SetLastLoginResponse,
	SetLastLoginInput,
	InternalSetLastLoginInput,
	InjectedSetLastLoginInput,
	UserInfoResponse,
	InternalUserInfoInput,
	InjectedUserInfoInput,
} from "./models";

export interface InternalClient {
	queries: {
		Hello: () => Promise<HelloResponse>;
		Messages: () => Promise<MessagesResponse>;
		MockQuery: () => Promise<MockQueryResponse>;
		UserInfo: (input: InternalUserInfoInput) => Promise<UserInfoResponse>;
	};
	mutations: {
		AddMessage: (input: InternalAddMessageInput) => Promise<AddMessageResponse>;
		ChangeUserName: (input: InternalChangeUserNameInput) => Promise<ChangeUserNameResponse>;
		DeleteAllMessagesByUserEmail: (
			input: InternalDeleteAllMessagesByUserEmailInput
		) => Promise<DeleteAllMessagesByUserEmailResponse>;
		SetLastLogin: (input: InternalSetLastLoginInput) => Promise<SetLastLoginResponse>;
	};
	withHeaders: (headers: { [key: string]: string }) => InternalClient;
}